pyproject.toml
[project]
name = "agentic-web-scraper-uao"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "crawl4ai>=0.8.0",
    "fastapi",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.2",
    "langchain-google-genai>=4.2.2",
    "langchain-ollama>=1.1.0",
    "langchain-text-splitters>=1.1.2",
    "ollama>=0.6.1",
    "playwright>=1.58.0",
    "streamlit>=1.56.0",
    "tavily-python>=0.7.24",
    "uvicorn[standard]",
]
